document-generation-solution-accelerator/src/requirements.txt
azure-identity==1.23.0
# Flask[async]==2.3.2
openai==1.84.0
azure-search-documents==11.6.0b9
azure-storage-blob==12.25.1
python-dotenv==1.1.0
azure-cosmos==4.9.0
azure-ai-projects==1.0.0b11
azure-ai-inference==1.0.0b9
quart==0.20.0
uvicorn==0.34.1
aiohttp==3.11.16
gunicorn==23.0.0
pydantic-settings==2.8.1
flake8==7.2.0
black==25.1.0
autoflake==2.3.1
isort==6.0.1
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-exporter-otlp-proto-grpc
azure-monitor-events-extension
opentelemetry-sdk==1.31.1
opentelemetry-api==1.31.1
opentelemetry-semantic-conventions==0.52b1
opentelemetry-instrumentation==0.52b1
azure-monitor-opentelemetry==1.6.8